requirements.txt
torch>=2.0.0
transformers>=4.30.0
flask==3.0.2
numpy>=1.21.0
scikit-learn>=1.0.0
openai==1.12.0
python-dotenv==1.0.1
gunicorn==21.2.0 